Mostly passing testcases except for column stuff
